refactor: schema modifiers
